fix main

+++ C0_Compiler/main.cpp
@@ -11,7 +11,7 @@ int main() {
 
 	cout << endl<<endl<<"Done. Press any key to quit" << endl;
 	clock_t end = clock();
-	double elapsed_secs = double(end - begin) / CLOCKS_PER_SEC;
+	double elapsed_secs = ((double)end - (double)begin) / CLOCKS_PER_SEC;
 	cout << elapsed_secs << endl;
 	getchar();
